delete alert renders

+++ public/js/post.js
@@ -1,19 +1,18 @@
 // delete post
 const delButtonHandler = async (event) => {
-    // if (event.target.hasAttribute('data-id')) {
-    //   const id = event.target.getAttribute('data-id');
-    const id = document.querySelector('#data-id');
+    if (event.target.hasAttribute('data-id')) {
+      const id = event.target.getAttribute('data-id');
 
       const response = await fetch(`/api/posts/${id}`, {
         method: 'DELETE',
       });
   
       if (response.ok) {
-        document.location.replace('../dashboard');
+        document.location.replace('/dashboard');
       } else {
         alert('Failed to delete post');
       }
-    // }
+    }
   };
 
 document.querySelector('.delete').addEventListener('click', delButtonHandler);
